// src/models/DB.interface.ts
export enum DB_TABLE {
    USERS = 'users',
    TODOLISTS = 'todo_lists'
}

export enum USERS_TABLE {
    USERID = 'user_id',
    NAME = 'name',
    EMAIL = 'email',
    PW = 'password',
}

export enum TODOSTABLE {
    ID = 'id',
    TITLE = 'title',
    DESCRIPT = 'description',
    USERID = 'user_id',
    CREATEDATE = 'create_date',
    UPDATEDATE = 'update_date'
}